Kinda working

+++ includes/types.h
@@ -6,7 +6,7 @@
 /*   By: acesar-l <acesar-l@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/24 19:16:31 by acesar-l          #+#    #+#             */
-/*   Updated: 2023/06/04 22:13:21 by acesar-l         ###   ########.fr       */
+/*   Updated: 2023/06/12 23:22:55 by acesar-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,10 @@ typedef enum e_bool
 
 typedef struct s_ray
 {
-	int		angle;
+	double	angle;
 	int		num_ray;
 	double	distance_to_wall;
-	double	orientation;
+	int		orientation;
 }	t_ray;
 
 typedef struct s_img
